requirements.txt
langchain==0.3.7
langchain-community==0.3.7
numpy==1.26.4
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.0
requests==2.31.0